RedshiftPipelineDemo/requirements.txt
-e .
aws_cdk.core
aws_cdk.aws_iam
aws_cdk.aws_ec2
aws_cdk.aws_redshift
aws_cdk.aws_glue
aws_cdk.aws_s3_assets
aws_cdk.aws_s3
aws_cdk.aws_s3_deployment